VB-4094 - update tests to move callGetSessions and callGetAvailableSessions.
