+++ src/test/kotlin/uk/gov/justice/digital/hmpps/visitscheduler/integration/admin/AdminGetSessionsTemplateLocationsWhenPrisonerIsTransitionalTest.kt
@@ -5,6 +5,7 @@ import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
+import org.springframework.http.HttpHeaders
 import org.springframework.http.HttpStatusCode
 import org.springframework.test.web.reactive.server.WebTestClient
 import org.springframework.test.web.reactive.server.WebTestClient.ResponseSpec
@@ -16,7 +17,7 @@ import uk.gov.justice.digital.hmpps.visitscheduler.dto.enums.TransitionalLocatio
 import uk.gov.justice.digital.hmpps.visitscheduler.dto.enums.TransitionalLocationTypes.ECL
 import uk.gov.justice.digital.hmpps.visitscheduler.dto.enums.TransitionalLocationTypes.RECP
 import uk.gov.justice.digital.hmpps.visitscheduler.dto.enums.TransitionalLocationTypes.TAP
-import uk.gov.justice.digital.hmpps.visitscheduler.dto.enums.UserType
+import uk.gov.justice.digital.hmpps.visitscheduler.dto.enums.UserType.STAFF
 import uk.gov.justice.digital.hmpps.visitscheduler.dto.sessions.VisitSessionDto
 import uk.gov.justice.digital.hmpps.visitscheduler.helper.AllowedSessionLocationHierarchy
 import uk.gov.justice.digital.hmpps.visitscheduler.integration.IntegrationTestBase
@@ -27,9 +28,11 @@ import uk.gov.justice.digital.hmpps.visitscheduler.model.entity.session.SessionT
 import java.time.LocalDate
 import java.time.LocalTime
 
-@DisplayName("Get  $ADMIN_SESSION_TEMPLATES_PATH with temp location")
+@DisplayName("GET $ADMIN_SESSION_TEMPLATES_PATH with temp location")
 class AdminGetSessionsTemplateLocationsWhenPrisonerIsTransitionalTest : IntegrationTestBase() {
   private val requiredRole = listOf("ROLE_VISIT_SCHEDULER")
+
+  private lateinit var authHttpHeaders: (HttpHeaders) -> Unit
 
   private val nextAllowedDay = getNextAllowedDay()
 
@@ -39,6 +42,7 @@ class AdminGetSessionsTemplateLocationsWhenPrisonerIsTransitionalTest : Integrat
   internal fun setUpTests() {
     prisonOther = prisonEntityHelper.create()
     prison = prisonEntityHelper.create("CR1")
+    authHttpHeaders = setAuthorisation(roles = requiredRole)
   }
 
   @Test
@@ -51,7 +55,7 @@ class AdminGetSessionsTemplateLocationsWhenPrisonerIsTransitionalTest : Integrat
     val sessionTemplate = setupSessionTemplate(prisonCode)
 
     // When
-    val responseSpec = callGetSessionsByPrisonerIdAndPrison(prisonCode, prisonerId)
+    val responseSpec = callGetSessions(prisonCode, prisonerId, userType = STAFF, authHttpHeaders = authHttpHeaders)
 
     // Then
     assertReturnedResult(responseSpec, sessionTemplate)
@@ -68,7 +72,7 @@ class AdminGetSessionsTemplateLocationsWhenPrisonerIsTransitionalTest : Integrat
     val sessionTemplate = setupSessionTemplate(prisonCode)
 
     // When
-    val responseSpec = callGetSessionsByPrisonerIdAndPrison(prisonCode, prisonerId)
+    val responseSpec = callGetSessions(prisonCode, prisonerId, userType = STAFF, authHttpHeaders = authHttpHeaders)
 
     // Then
     assertReturnedResult(responseSpec, sessionTemplate)
@@ -86,7 +90,7 @@ class AdminGetSessionsTemplateLocationsWhenPrisonerIsTransitionalTest : Integrat
     val sessionTemplateTAP = setupSessionTemplate(prisonCode, allowedPermittedLocations = listOf(allowedSessionLocationHierarchy))
 
     // When
-    val responseSpec = callGetSessionsByPrisonerIdAndPrison(prisonCode, prisonerId)
+    val responseSpec = callGetSessions(prisonCode, prisonerId, userType = STAFF, authHttpHeaders = authHttpHeaders)
 
     // Then
     assertReturnedResult(responseSpec, sessionTemplateTAP)
@@ -103,7 +107,7 @@ class AdminGetSessionsTemplateLocationsWhenPrisonerIsTransitionalTest : Integrat
     val sessionTemplate = setupSessionTemplate(prisonCode)
 
     // When
-    val responseSpec = callGetSessionsByPrisonerIdAndPrison(prisonCode, prisonerId)
+    val responseSpec = callGetSessions(prisonCode, prisonerId, userType = STAFF, authHttpHeaders = authHttpHeaders)
 
     // Then
     assertReturnedResult(responseSpec, sessionTemplate)
@@ -120,7 +124,7 @@ class AdminGetSessionsTemplateLocationsWhenPrisonerIsTransitionalTest : Integrat
     val sessionTemplate = setupSessionTemplate(prisonCode)
 
     // When
-    val responseSpec = callGetSessionsByPrisonerIdAndPrison(prisonCode, prisonerId)
+    val responseSpec = callGetSessions(prisonCode, prisonerId, userType = STAFF, authHttpHeaders = authHttpHeaders)
 
     // Then
     assertReturnedResult(responseSpec, sessionTemplate)
@@ -137,7 +141,7 @@ class AdminGetSessionsTemplateLocationsWhenPrisonerIsTransitionalTest : Integrat
     val sessionTemplate = setupSessionTemplate(prisonCode)
 
     // When
-    val responseSpec = callGetSessionsByPrisonerIdAndPrison(prisonCode, prisonerId)
+    val responseSpec = callGetSessions(prisonCode, prisonerId, userType = STAFF, authHttpHeaders = authHttpHeaders)
 
     // Then
     assertReturnedResult(responseSpec, sessionTemplate)
@@ -155,7 +159,7 @@ class AdminGetSessionsTemplateLocationsWhenPrisonerIsTransitionalTest : Integrat
     val sessionTemplate = setupSessionTemplate(prisonCode)
 
     // When
-    val responseSpec = callGetSessionsByPrisonerIdAndPrison(prisonCode, prisonerId)
+    val responseSpec = callGetSessions(prisonCode, prisonerId, userType = STAFF, authHttpHeaders = authHttpHeaders)
 
     // Then
     assertReturnedResult(responseSpec, sessionTemplate)
@@ -172,7 +176,7 @@ class AdminGetSessionsTemplateLocationsWhenPrisonerIsTransitionalTest : Integrat
     setupSessionTemplate(prisonCode)
 
     // When
-    val responseSpec = callGetSessionsByPrisonerIdAndPrison(prisonCode, prisonerId)
+    val responseSpec = callGetSessions(prisonCode, prisonerId, userType = STAFF, authHttpHeaders = authHttpHeaders)
 
     // Then
     assertNoResult(responseSpec)
@@ -189,7 +193,7 @@ class AdminGetSessionsTemplateLocationsWhenPrisonerIsTransitionalTest : Integrat
     prisonOffenderSearchMockServer.stubGetPrisonerByString(prisonerId = prisonerId, prisonCode = prisonCode)
 
     // When
-    val responseSpec = callGetSessionsByPrisonerIdAndPrison(prisonCode, prisonerId)
+    val responseSpec = callGetSessions(prisonCode, prisonerId, userType = STAFF, authHttpHeaders = authHttpHeaders)
 
     // Then
     assertNoResult(responseSpec)
@@ -208,7 +212,7 @@ class AdminGetSessionsTemplateLocationsWhenPrisonerIsTransitionalTest : Integrat
     prisonOffenderSearchMockServer.stubGetPrisoner(prisonerId = prisonerId, null)
 
     // When
-    val responseSpec = callGetSessionsByPrisonerIdAndPrison(prisonCode, prisonerId)
+    val responseSpec = callGetSessions(prisonCode, prisonerId, userType = STAFF, authHttpHeaders = authHttpHeaders)
 
     // Then
     assertErrorResult(
@@ -231,7 +235,7 @@ class AdminGetSessionsTemplateLocationsWhenPrisonerIsTransitionalTest : Integrat
     prisonOffenderSearchMockServer.stubGetPrisoner(prisonerId, null)
 
     // When
-    val responseSpec = callGetSessionsByPrisonerIdAndPrison(prisonCode, prisonerId)
+    val responseSpec = callGetSessions(prisonCode, prisonerId, userType = STAFF, authHttpHeaders = authHttpHeaders)
 
     // Then
     assertErrorResult(
@@ -288,14 +292,6 @@ class AdminGetSessionsTemplateLocationsWhenPrisonerIsTransitionalTest : Integrat
       includeLocationGroupType = includeLocationGroupType,
     )
   }
-
-  private fun callGetSessionsByPrisonerIdAndPrison(
-    prisonId: String,
-    prisonerId: String,
-    userType: UserType = UserType.STAFF,
-  ): ResponseSpec = webTestClient.get().uri("/visit-sessions?prisonId=$prisonId&prisonerId=$prisonerId&userType=$userType")
-    .headers(setAuthorisation(roles = requiredRole))
-    .exchange()
 
   private fun getNextAllowedDay(): LocalDate {
     // The two days is based on the default SessionService.policyNoticeDaysMin
