+++ src/test/kotlin/uk/gov/justice/digital/hmpps/visitscheduler/integration/session/GetAvailableSessionsTest.kt
@@ -4,6 +4,7 @@ import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
+import org.springframework.http.HttpHeaders
 import org.springframework.http.HttpStatus.BAD_REQUEST
 import org.springframework.test.web.reactive.server.WebTestClient.BodyContentSpec
 import org.springframework.test.web.reactive.server.WebTestClient.ResponseSpec
@@ -23,14 +24,13 @@ import uk.gov.justice.digital.hmpps.visitscheduler.dto.sessions.AvailableVisitSe
 import uk.gov.justice.digital.hmpps.visitscheduler.helper.submitApplication
 import uk.gov.justice.digital.hmpps.visitscheduler.integration.IntegrationTestBase
 import uk.gov.justice.digital.hmpps.visitscheduler.model.entity.session.SessionTemplate
-import uk.gov.justice.digital.hmpps.visitscheduler.service.DateRange
 import java.time.DayOfWeek.MONDAY
 import java.time.DayOfWeek.SUNDAY
 import java.time.LocalDate
 import java.time.LocalTime
 import java.time.temporal.TemporalAdjusters
 
-@DisplayName("Get /visit-sessions/available")
+@DisplayName("GET $VISIT_SESSIONS_AVAILABLE_CONTROLLER_PATH")
 class GetAvailableSessionsTest : IntegrationTestBase() {
 
   private val requiredRole = listOf("ROLE_VISIT_SCHEDULER")
@@ -39,8 +39,11 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
 
   private val prisonCode = "STC"
 
+  private lateinit var authHttpHeaders: (HttpHeaders) -> Unit
+
   @BeforeEach
   internal fun setUpTests() {
+    authHttpHeaders = setAuthorisation(roles = requiredRole)
     prison = prisonEntityHelper.create(prisonCode = prisonCode)
     prisonOffenderSearchMockServer.stubGetPrisonerByString(prisonerId, prisonCode)
   }
@@ -66,6 +69,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode,
       prisonerId,
       OPEN,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -108,6 +112,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode,
       prisonerId,
       CLOSED,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -144,7 +149,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
     )
 
     // When
-    val responseSpec = callGetAvailableSessions(prisonCode, prisonerId, OPEN, 2, 28, excludedApplicationReference = reservedApplication.reference)
+    val responseSpec = callGetAvailableSessions(prisonCode, prisonerId, OPEN, 2, 28, excludedApplicationReference = reservedApplication.reference, authHttpHeaders = authHttpHeaders)
 
     // Then
     responseSpec.expectStatus().isOk
@@ -181,7 +186,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
     )
 
     // When
-    val responseSpec = callGetAvailableSessions(prisonCode, prisonerId, OPEN, 1, 7, excludedApplicationReference = reservedApplication.reference)
+    val responseSpec = callGetAvailableSessions(prisonCode, prisonerId, OPEN, 1, 7, excludedApplicationReference = reservedApplication.reference, authHttpHeaders = authHttpHeaders)
 
     // Then
     responseSpec.expectStatus().isOk
@@ -219,7 +224,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
     )
 
     // When
-    val responseSpec = callGetAvailableSessions(prisonCode, prisonerId, OPEN, 1, 7)
+    val responseSpec = callGetAvailableSessions(prisonCode, prisonerId, OPEN, 1, 7, authHttpHeaders = authHttpHeaders)
 
     // Then
     responseSpec.expectStatus().isOk
@@ -255,7 +260,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
     )
 
     // When
-    val responseSpec = callGetAvailableSessions(prisonCode, prisonerId, CLOSED, 2, 28, excludedApplicationReference = reservedApplication.reference)
+    val responseSpec = callGetAvailableSessions(prisonCode, prisonerId, CLOSED, 2, 28, excludedApplicationReference = reservedApplication.reference, authHttpHeaders = authHttpHeaders)
 
     // Then
     responseSpec.expectStatus().isOk
@@ -293,7 +298,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
     )
 
     // When
-    val responseSpec = callGetAvailableSessions(prisonCode, prisonerId, CLOSED, 1, 7, excludedApplicationReference = reservedApplication.reference)
+    val responseSpec = callGetAvailableSessions(prisonCode, prisonerId, CLOSED, 1, 7, excludedApplicationReference = reservedApplication.reference, authHttpHeaders = authHttpHeaders)
 
     // Then
     responseSpec.expectStatus().isOk
@@ -331,7 +336,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
     )
 
     // When
-    val responseSpec = callGetAvailableSessions(prisonCode, prisonerId, CLOSED, 1, 7)
+    val responseSpec = callGetAvailableSessions(prisonCode, prisonerId, CLOSED, 1, 7, authHttpHeaders = authHttpHeaders)
 
     // Then
     responseSpec.expectStatus().isOk
@@ -380,6 +385,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       CLOSED,
       2,
       28,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -438,6 +444,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       CLOSED,
       2,
       28,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -507,6 +514,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       2,
       28,
       username = "username",
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -566,6 +574,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       CLOSED,
       2,
       28,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -617,6 +626,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       OPEN,
       2,
       28,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -667,6 +677,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode,
       prisonerId,
       OPEN,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -735,6 +746,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode,
       prisonerId,
       OPEN,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -806,6 +818,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode,
       prisonerId,
       OPEN,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -843,6 +856,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       OPEN,
       policyNoticeDaysMin = policyNoticeDaysMin,
       policyNoticeDaysMax = policyNoticeDaysMax,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -877,6 +891,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode = "AWE",
       prisonerId,
       OPEN,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -938,6 +953,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode,
       prisonerId,
       OPEN,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -980,6 +996,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode,
       prisonerId,
       OPEN,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -1016,6 +1033,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode,
       prisonerId,
       OPEN,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -1051,6 +1069,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode,
       prisonerId,
       OPEN,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -1090,6 +1109,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode,
       prisonerId,
       OPEN,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -1129,6 +1149,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode,
       prisonerId,
       OPEN,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -1172,6 +1193,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode,
       prisonerId,
       OPEN,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -1215,6 +1237,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode,
       prisonerId,
       OPEN,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -1247,6 +1270,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode,
       prisonerId,
       OPEN,
+      authHttpHeaders = authHttpHeaders,
     )
     // Then
     val returnResult = responseSpec.expectStatus().isOk.expectBody()
@@ -1296,6 +1320,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode,
       prisonerId,
       OPEN,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -1345,6 +1370,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode,
       prisonerId,
       OPEN,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -1382,6 +1408,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode,
       prisonerId,
       OPEN,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -1421,6 +1448,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       OPEN,
       0,
       28,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -1458,6 +1486,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode,
       prisonerId,
       OPEN,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -1490,6 +1519,8 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode,
       prisonerId,
       OPEN,
+
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -1516,6 +1547,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       OPEN,
       policyNoticeDaysMin,
       policyNoticeDaysMax,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -1540,6 +1572,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       OPEN,
       policyNoticeDaysMin,
       policyNoticeDaysMax,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -1563,6 +1596,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       OPEN,
       policyNoticeDaysMin,
       policyNoticeDaysMax,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -1582,6 +1616,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode,
       prisonerId,
       OPEN,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -1601,6 +1636,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode,
       prisonerId,
       OPEN,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -1682,6 +1718,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode,
       prisonerId,
       OPEN,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -1736,6 +1773,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode,
       prisonerId,
       OPEN,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -1834,6 +1872,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode,
       prisonerId,
       OPEN,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -1981,6 +2020,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode,
       prisonerId,
       OPEN,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -2006,6 +2046,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode,
       prisonerId,
       OPEN,
+      authHttpHeaders = authHttpHeaders,
     )
     // Then
     assertResponseLength(responseSpec, 4)
@@ -2032,6 +2073,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode,
       prisonerId,
       OPEN,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -2072,6 +2114,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode,
       prisonerId,
       OPEN,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -2147,6 +2190,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode,
       prisonerId,
       OPEN,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -2191,6 +2235,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode,
       prisonerId,
       OPEN,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -2232,6 +2277,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode,
       prisonerId,
       OPEN,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -2272,6 +2318,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode,
       prisonerId,
       OPEN,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -2311,6 +2358,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode,
       prisonerId,
       OPEN,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -2341,6 +2389,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonerId = prisonerId,
       sessionRestriction = OPEN,
       username = createdByUser,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -2374,6 +2423,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonerId = prisonerId,
       sessionRestriction = OPEN,
       username = "other-user",
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -2406,6 +2456,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode = prisonCode,
       prisonerId = prisonerId,
       sessionRestriction = OPEN,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -2431,6 +2482,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       incorrectPrisonCode,
       prisonerId,
       OPEN,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -2470,6 +2522,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode,
       prisonerId,
       OPEN,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -2508,6 +2561,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode,
       prisonerId,
       OPEN,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -2546,6 +2600,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode,
       prisonerId,
       OPEN,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -2582,6 +2637,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
       prisonCode,
       prisonerId,
       OPEN,
+      authHttpHeaders = authHttpHeaders,
     )
 
     // Then
@@ -2609,7 +2665,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
     )
 
     // When
-    val responseSpec = callGetAvailableSessions(prisonCode = "AWE", prisonerId = prisonerId, sessionRestriction = OPEN)
+    val responseSpec = callGetAvailableSessions(prisonCode = "AWE", prisonerId = prisonerId, sessionRestriction = OPEN, authHttpHeaders = authHttpHeaders)
 
     // Then
     val returnResult = responseSpec.expectStatus().isOk
@@ -2652,7 +2708,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
     )
 
     // When
-    val responseSpec = callGetAvailableSessions(prisonCode = "AWE", prisonerId = prisonerId, sessionRestriction = OPEN)
+    val responseSpec = callGetAvailableSessions(prisonCode = "AWE", prisonerId = prisonerId, sessionRestriction = OPEN, authHttpHeaders = authHttpHeaders)
 
     // Then
     val returnResult = responseSpec.expectStatus().isOk
@@ -2699,7 +2755,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
     )
 
     // When
-    val responseSpec = callGetAvailableSessions(prisonCode = "AWE", prisonerId = prisonerId, OPEN)
+    val responseSpec = callGetAvailableSessions(prisonCode = "AWE", prisonerId = prisonerId, OPEN, authHttpHeaders = authHttpHeaders)
 
     // Then
     val returnResult = responseSpec.expectStatus().isOk
@@ -2747,7 +2803,7 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
     )
 
     // When
-    val responseSpec = callGetAvailableSessions(prisonCode = "AWE", prisonerId = prisonerId, sessionRestriction = OPEN)
+    val responseSpec = callGetAvailableSessions(prisonCode = "AWE", prisonerId = prisonerId, sessionRestriction = OPEN, authHttpHeaders = authHttpHeaders)
 
     // Then
     val returnResult = responseSpec.expectStatus().isOk
@@ -2760,70 +2816,6 @@ class GetAvailableSessionsTest : IntegrationTestBase() {
     assertSession(visitSessionResults[1], nextWeek, sessionTemplate2, OPEN)
   }
 
-  private fun callGetAvailableSessions(
-    prisonCode: String? = "SPC",
-    prisonerId: String,
-    sessionRestriction: SessionRestriction,
-    policyNoticeDaysMin: Int = 2,
-    policyNoticeDaysMax: Int = 28,
-    excludedApplicationReference: String? = null,
-    username: String? = null,
-    userType: UserType = UserType.PUBLIC,
-  ): ResponseSpec {
-    val today = LocalDate.now()
-    val fromDate = today.plusDays(policyNoticeDaysMin.toLong())
-    val toDate = today.plusDays(policyNoticeDaysMax.toLong())
-    val dateRange = DateRange(fromDate, toDate)
-
-    val uri = VISIT_SESSIONS_AVAILABLE_CONTROLLER_PATH
-    val uriQueryParams = getAvailableSessionsQueryParams(
-      prisonCode = prisonCode!!,
-      prisonerId = prisonerId,
-      sessionRestriction = sessionRestriction,
-      dateRange = dateRange,
-      policyNoticeDaysMin = policyNoticeDaysMin,
-      policyNoticeDaysMax = policyNoticeDaysMax,
-      excludedApplicationReference = excludedApplicationReference,
-      username = username,
-      userType = userType,
-    ).joinToString("&")
-
-    return webTestClient.get().uri("$uri?$uriQueryParams")
-      .headers(setAuthorisation(roles = requiredRole))
-      .exchange()
-  }
-
-  private fun getAvailableSessionsQueryParams(
-    prisonCode: String,
-    prisonerId: String,
-    sessionRestriction: SessionRestriction,
-    dateRange: DateRange,
-    policyNoticeDaysMin: Int,
-    policyNoticeDaysMax: Int,
-    excludedApplicationReference: String?,
-    username: String?,
-    userType: UserType,
-  ): List<String> {
-    val queryParams = ArrayList<String>()
-    queryParams.add("prisonId=$prisonCode")
-    queryParams.add("prisonerId=$prisonerId")
-    queryParams.add("sessionRestriction=$sessionRestriction")
-    queryParams.add("fromDate=${dateRange.fromDate}")
-    queryParams.add("toDate=${dateRange.toDate}")
-    queryParams.add("policyNoticeDaysMin=$policyNoticeDaysMin")
-    queryParams.add("policyNoticeDaysMax=$policyNoticeDaysMax")
-
-    excludedApplicationReference?.let {
-      queryParams.add("excludedApplicationReference=$excludedApplicationReference")
-    }
-
-    username?.let {
-      queryParams.add("username=$username")
-    }
-
-    queryParams.add("userType=${userType.name}")
-    return queryParams
-  }
   private fun getNextAllowedDay(): LocalDate {
     // The 3 days is based on the default SessionService.policyNoticeDaysMin
     return LocalDate.now().plusDays(3)
