+++ src/test/kotlin/uk/gov/justice/digital/hmpps/visitscheduler/integration/session/GetSessionsDoubleBookingTest.kt
@@ -5,16 +5,16 @@ import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
 import org.springframework.test.web.reactive.server.WebTestClient.BodyContentSpec
-import org.springframework.test.web.reactive.server.WebTestClient.ResponseSpec
+import uk.gov.justice.digital.hmpps.visitscheduler.controller.VISIT_SESSION_CONTROLLER_PATH
 import uk.gov.justice.digital.hmpps.visitscheduler.dto.enums.SessionConflict.DOUBLE_BOOKING_OR_RESERVATION
-import uk.gov.justice.digital.hmpps.visitscheduler.dto.enums.UserType
+import uk.gov.justice.digital.hmpps.visitscheduler.dto.enums.UserType.STAFF
 import uk.gov.justice.digital.hmpps.visitscheduler.dto.enums.VisitStatus
 import uk.gov.justice.digital.hmpps.visitscheduler.dto.sessions.VisitSessionDto
 import uk.gov.justice.digital.hmpps.visitscheduler.integration.IntegrationTestBase
 import uk.gov.justice.digital.hmpps.visitscheduler.model.entity.session.SessionTemplate
 import java.time.LocalDate
 
-@DisplayName("Get /visit-sessions - Tests to check for DOUBLE_BOOKING_OR_RESERVATION flag.")
+@DisplayName("GET $VISIT_SESSION_CONTROLLER_PATH - Tests to check for DOUBLE_BOOKING_OR_RESERVATION flag.")
 class GetSessionsDoubleBookingTest : IntegrationTestBase() {
 
   private val requiredRole = listOf("ROLE_VISIT_SCHEDULER")
@@ -27,8 +27,11 @@ class GetSessionsDoubleBookingTest : IntegrationTestBase() {
 
   private lateinit var visitDate: LocalDate
 
+  private lateinit var authHttpHeaders: (org.springframework.http.HttpHeaders) -> Unit
+
   @BeforeEach
   internal fun setUpTests() {
+    authHttpHeaders = setAuthorisation(roles = requiredRole)
     visitDate = getNextAllowedDay()
     prison = prisonEntityHelper.create(prisonCode = prisonCode)
     prisonOffenderSearchMockServer.stubGetPrisonerByString(prisonerId, prisonCode)
@@ -55,7 +58,7 @@ class GetSessionsDoubleBookingTest : IntegrationTestBase() {
     )
 
     // When
-    val responseSpec = callGetSessions(prisonCode, prisonerId)
+    val responseSpec = callGetSessions(prisonCode, prisonerId, userType = STAFF, authHttpHeaders = authHttpHeaders)
 
     // Then
     val returnResult = responseSpec.expectStatus().isOk.expectBody()
@@ -78,7 +81,7 @@ class GetSessionsDoubleBookingTest : IntegrationTestBase() {
     )
 
     // When
-    val responseSpec = callGetSessions(prisonCode, prisonerId, username = null)
+    val responseSpec = callGetSessions(prisonCode, prisonerId, userName = null, userType = STAFF, authHttpHeaders = authHttpHeaders)
 
     // Then
     val returnResult = responseSpec.expectStatus().isOk.expectBody()
@@ -103,7 +106,7 @@ class GetSessionsDoubleBookingTest : IntegrationTestBase() {
     )
 
     // When
-    val responseSpec = callGetSessions(prisonCode, prisonerId, username = currentUser)
+    val responseSpec = callGetSessions(prisonCode, prisonerId, userName = currentUser, userType = STAFF, authHttpHeaders = authHttpHeaders)
 
     // Then
     val returnResult = responseSpec.expectStatus().isOk.expectBody()
@@ -128,38 +131,13 @@ class GetSessionsDoubleBookingTest : IntegrationTestBase() {
     )
 
     // When
-    val responseSpec = callGetSessions(prisonCode, prisonerId, username = currentUser)
+    val responseSpec = callGetSessions(prisonCode, prisonerId, userName = currentUser, userType = STAFF, authHttpHeaders = authHttpHeaders)
 
     // Then
     val returnResult = responseSpec.expectStatus().isOk.expectBody()
     val visitSessionResults = getResults(returnResult)
     assertThat(visitSessionResults.size).isEqualTo(1)
     assertThat(visitSessionResults[0].sessionConflicts.isEmpty())
-  }
-
-  private fun callGetSessions(
-    prisonCode: String,
-    prisonerId: String,
-    policyNoticeDaysMin: Int = 2,
-    policyNoticeDaysMax: Int = 28,
-    username: String? = null,
-    userType: UserType = UserType.STAFF,
-  ): ResponseSpec {
-    val uriQueryParams = mutableListOf(
-      "prisonId=$prisonCode",
-      "prisonerId=$prisonerId",
-      "min=$policyNoticeDaysMin",
-      "max=$policyNoticeDaysMax",
-      "userType=$userType",
-    ).also { params ->
-      username?.let {
-        params.add("username=$username")
-      }
-    }.joinToString("&")
-
-    return webTestClient.get().uri("/visit-sessions?$uriQueryParams")
-      .headers(setAuthorisation(roles = requiredRole))
-      .exchange()
   }
 
   private fun getNextAllowedDay(): LocalDate = LocalDate.now().plusDays(3)
