+++ src/test/kotlin/uk/gov/justice/digital/hmpps/visitscheduler/integration/session/GetSessionsWithExcludeLocationsTest.kt
@@ -4,9 +4,11 @@ import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
+import org.springframework.http.HttpHeaders
 import org.springframework.test.web.reactive.server.WebTestClient
+import uk.gov.justice.digital.hmpps.visitscheduler.controller.VISIT_SESSION_CONTROLLER_PATH
 import uk.gov.justice.digital.hmpps.visitscheduler.dto.enums.TransitionalLocationTypes
-import uk.gov.justice.digital.hmpps.visitscheduler.dto.enums.UserType
+import uk.gov.justice.digital.hmpps.visitscheduler.dto.enums.UserType.STAFF
 import uk.gov.justice.digital.hmpps.visitscheduler.dto.sessions.VisitSessionDto
 import uk.gov.justice.digital.hmpps.visitscheduler.helper.AllowedSessionLocationHierarchy
 import uk.gov.justice.digital.hmpps.visitscheduler.integration.IntegrationTestBase
@@ -15,8 +17,10 @@ import uk.gov.justice.digital.hmpps.visitscheduler.model.entity.session.location
 import java.time.LocalDate
 import java.time.LocalTime
 
-@DisplayName("Get /visit-sessions - tests for exclude locations")
+@DisplayName("GET $VISIT_SESSION_CONTROLLER_PATH - tests for exclude locations")
 class GetSessionsWithExcludeLocationsTest : IntegrationTestBase() {
+  private lateinit var authHttpHeaders: (HttpHeaders) -> Unit
+
   private val requiredRole = listOf("ROLE_VISIT_SCHEDULER")
 
   private val nextAllowedDay = getNextAllowedDay()
@@ -30,6 +34,8 @@ class GetSessionsWithExcludeLocationsTest : IntegrationTestBase() {
 
   @BeforeEach
   internal fun createAllSessionTemplates() {
+    authHttpHeaders = setAuthorisation(roles = requiredRole)
+
     prison = prisonEntityHelper.create("SWL")
     tapAsLocation = sessionLocationGroupHelper.create(
       prisonCode = prison.code,
@@ -147,7 +153,7 @@ class GetSessionsWithExcludeLocationsTest : IntegrationTestBase() {
     prisonApiMockServer.stubGetPrisonerHousingLocation(prisonerId, prisonerInternalLocation)
 
     // When
-    val responseSpec = callGetSessionsByPrisonerIdAndPrison(prison.code, prisonerId)
+    val responseSpec = callGetSessions(prisonCode = prison.code, prisonerId = prisonerId, userType = STAFF, authHttpHeaders = authHttpHeaders)
 
     // Then
     val returnResult = responseSpec.expectStatus().isOk.expectBody()
@@ -167,7 +173,7 @@ class GetSessionsWithExcludeLocationsTest : IntegrationTestBase() {
     prisonApiMockServer.stubGetPrisonerHousingLocation(prisonerId, prisonerInternalLocation)
 
     // When
-    val responseSpec = callGetSessionsByPrisonerIdAndPrison(prison.code, prisonerId)
+    val responseSpec = callGetSessions(prisonCode = prison.code, prisonerId = prisonerId, userType = STAFF, authHttpHeaders = authHttpHeaders)
 
     // Then
     val returnResult = responseSpec.expectStatus().isOk.expectBody()
@@ -189,7 +195,7 @@ class GetSessionsWithExcludeLocationsTest : IntegrationTestBase() {
     prisonApiMockServer.stubGetPrisonerHousingLocation(prisonerId, prisonerInternalLocation)
 
     // When
-    val responseSpec = callGetSessionsByPrisonerIdAndPrison(prison.code, prisonerId)
+    val responseSpec = callGetSessions(prisonCode = prison.code, prisonerId = prisonerId, userType = STAFF, authHttpHeaders = authHttpHeaders)
 
     // Then
     val returnResult = responseSpec.expectStatus().isOk.expectBody()
@@ -213,7 +219,7 @@ class GetSessionsWithExcludeLocationsTest : IntegrationTestBase() {
     prisonApiMockServer.stubGetPrisonerHousingLocation(prisonerId, prisonerInternalLocation)
 
     // When
-    val responseSpec = callGetSessionsByPrisonerIdAndPrison(prison.code, prisonerId)
+    val responseSpec = callGetSessions(prisonCode = prison.code, prisonerId = prisonerId, userType = STAFF, authHttpHeaders = authHttpHeaders)
 
     // Then
     val returnResult = responseSpec.expectStatus().isOk.expectBody()
@@ -234,7 +240,7 @@ class GetSessionsWithExcludeLocationsTest : IntegrationTestBase() {
     prisonApiMockServer.stubGetPrisonerHousingLocation(prisonerId, prisonerInternalLocation)
 
     // When
-    val responseSpec = callGetSessionsByPrisonerIdAndPrison(prison.code, prisonerId)
+    val responseSpec = callGetSessions(prisonCode = prison.code, prisonerId = prisonerId, userType = STAFF, authHttpHeaders = authHttpHeaders)
 
     // Then
     val returnResult = responseSpec.expectStatus().isOk.expectBody()
@@ -256,7 +262,7 @@ class GetSessionsWithExcludeLocationsTest : IntegrationTestBase() {
     prisonApiMockServer.stubGetPrisonerHousingLocation(prisonerId, prisonerInternalLocation)
 
     // When
-    val responseSpec = callGetSessionsByPrisonerIdAndPrison(prison.code, prisonerId)
+    val responseSpec = callGetSessions(prisonCode = prison.code, prisonerId = prisonerId, userType = STAFF, authHttpHeaders = authHttpHeaders)
 
     // Then
     val returnResult = responseSpec.expectStatus().isOk.expectBody()
@@ -280,7 +286,7 @@ class GetSessionsWithExcludeLocationsTest : IntegrationTestBase() {
     prisonApiMockServer.stubGetPrisonerHousingLocation(prisonerId, prisonerInternalLocation)
 
     // When
-    val responseSpec = callGetSessionsByPrisonerIdAndPrison(prison.code, prisonerId)
+    val responseSpec = callGetSessions(prisonCode = prison.code, prisonerId = prisonerId, userType = STAFF, authHttpHeaders = authHttpHeaders)
 
     // Then
     val returnResult = responseSpec.expectStatus().isOk.expectBody()
@@ -304,7 +310,7 @@ class GetSessionsWithExcludeLocationsTest : IntegrationTestBase() {
     prisonApiMockServer.stubGetPrisonerHousingLocation(prisonerId, prisonerInternalLocation)
 
     // When
-    val responseSpec = callGetSessionsByPrisonerIdAndPrison(prison.code, prisonerId)
+    val responseSpec = callGetSessions(prisonCode = prison.code, prisonerId = prisonerId, userType = STAFF, authHttpHeaders = authHttpHeaders)
 
     // Then
     val returnResult = responseSpec.expectStatus().isOk.expectBody()
@@ -325,7 +331,7 @@ class GetSessionsWithExcludeLocationsTest : IntegrationTestBase() {
     prisonApiMockServer.stubGetPrisonerHousingLocation(prisonerId, prisonerInternalLocation)
 
     // When
-    val responseSpec = callGetSessionsByPrisonerIdAndPrison(prison.code, prisonerId)
+    val responseSpec = callGetSessions(prisonCode = prison.code, prisonerId = prisonerId, userType = STAFF, authHttpHeaders = authHttpHeaders)
 
     // Then
     val returnResult = responseSpec.expectStatus().isOk.expectBody()
@@ -346,7 +352,7 @@ class GetSessionsWithExcludeLocationsTest : IntegrationTestBase() {
     prisonApiMockServer.stubGetPrisonerHousingLocation(prisonerId, prisonerInternalLocation)
 
     // When
-    val responseSpec = callGetSessionsByPrisonerIdAndPrison(prison.code, prisonerId)
+    val responseSpec = callGetSessions(prisonCode = prison.code, prisonerId = prisonerId, userType = STAFF, authHttpHeaders = authHttpHeaders)
 
     // Then
     val returnResult = responseSpec.expectStatus().isOk.expectBody()
@@ -372,7 +378,7 @@ class GetSessionsWithExcludeLocationsTest : IntegrationTestBase() {
     prisonApiMockServer.stubGetPrisonerHousingLocation(prisonerId, prisonerInternalLocation, prisonerTemporaryLocation)
 
     // When
-    val responseSpec = callGetSessionsByPrisonerIdAndPrison(prison.code, prisonerId)
+    val responseSpec = callGetSessions(prisonCode = prison.code, prisonerId = prisonerId, userType = STAFF, authHttpHeaders = authHttpHeaders)
 
     // Then
     val returnResult = responseSpec.expectStatus().isOk.expectBody()
@@ -398,7 +404,7 @@ class GetSessionsWithExcludeLocationsTest : IntegrationTestBase() {
     prisonApiMockServer.stubGetPrisonerHousingLocation(offenderNo = prisonerId, internalLocation = prisonerTemporaryLocation, lastPermanentLevels = prisonerInternalLocation)
 
     // When
-    val responseSpec = callGetSessionsByPrisonerIdAndPrison(prison.code, prisonerId)
+    val responseSpec = callGetSessions(prisonCode = prison.code, prisonerId = prisonerId, userType = STAFF, authHttpHeaders = authHttpHeaders)
 
     // Then
     val returnResult = responseSpec.expectStatus().isOk.expectBody()
@@ -436,7 +442,7 @@ class GetSessionsWithExcludeLocationsTest : IntegrationTestBase() {
     prisonApiMockServer.stubGetPrisonerHousingLocation(offenderNo = prisonerId, internalLocation = prisonerTemporaryLocation, lastPermanentLevels = prisonerInternalLocation)
 
     // When
-    val responseSpec = callGetSessionsByPrisonerIdAndPrison(prison.code, prisonerId)
+    val responseSpec = callGetSessions(prisonCode = prison.code, prisonerId = prisonerId, userType = STAFF, authHttpHeaders = authHttpHeaders)
 
     // Then
     val returnResult = responseSpec.expectStatus().isOk.expectBody()
@@ -447,14 +453,6 @@ class GetSessionsWithExcludeLocationsTest : IntegrationTestBase() {
     // only TAP session is available to that prisoner
     assertSession(visitSessionResults[0], nextAllowedDay, tapSessionTemplate)
   }
-
-  private fun callGetSessionsByPrisonerIdAndPrison(
-    prisonId: String,
-    prisonerId: String,
-    userType: UserType = UserType.STAFF,
-  ): WebTestClient.ResponseSpec = webTestClient.get().uri("/visit-sessions?prisonId=$prisonId&prisonerId=$prisonerId&userType=$userType")
-    .headers(setAuthorisation(roles = requiredRole))
-    .exchange()
 
   private fun getNextAllowedDay(): LocalDate {
     // The two days is based on the default SessionService.policyNoticeDaysMin
